更新bmobsdk

+++ app/src/main/java/com/yuri/dreamlinkcost/model/Main.java
@@ -5,9 +5,6 @@ import android.os.Bundle;
 import android.os.Message;
 
 import com.activeandroid.query.Select;
-import com.bmob.BmobProFile;
-import com.bmob.btp.callback.DownloadListener;
-import com.bmob.btp.callback.UploadListener;
 import com.yuri.dreamlinkcost.BuildConfig;
 import com.yuri.dreamlinkcost.R;
 import com.yuri.dreamlinkcost.bean.Bmob.BmobCost;
@@ -15,7 +12,6 @@ import com.yuri.dreamlinkcost.bean.Bmob.Version;
 import com.yuri.dreamlinkcost.bean.table.Cost;
 import com.yuri.dreamlinkcost.bean.table.Title;
 import com.yuri.dreamlinkcost.model.impl.IMain;
-import com.yuri.dreamlinkcost.utils.CommonUtils;
 import com.yuri.dreamlinkcost.utils.NetUtil;
 import com.yuri.dreamlinkcost.utils.SharedPreferencesUtil;
 import com.yuri.xlog.Log;
@@ -24,9 +20,13 @@ import java.io.File;
 import java.util.List;
 
 import cn.bmob.v3.BmobQuery;
+import cn.bmob.v3.datatype.BmobFile;
+import cn.bmob.v3.exception.BmobException;
+import cn.bmob.v3.listener.DownloadFileListener;
 import cn.bmob.v3.listener.FindListener;
 import cn.bmob.v3.listener.SaveListener;
 import cn.bmob.v3.listener.UpdateListener;
+import cn.bmob.v3.listener.UploadFileListener;
 
 /**
  * Created by Yuri on 2016/1/16.
@@ -67,99 +67,99 @@ public class Main extends BaseMain implements IMain {
             }
         }
         BmobQuery<Version> bmobQuery = new BmobQuery();
-        bmobQuery.findObjects(context, new FindListener<Version>() {
+        bmobQuery.findObjects(new FindListener<Version>() {
             @Override
-            public void onSuccess(List<Version> list) {
-                if (list != null && list.size() > 0) {
-                    int serverVersionCode = list.get(0).version_code;
-                    int currentVersionCode = BuildConfig.VERSION_CODE;
-                    Log.d("serverVersionCode:" + serverVersionCode + ",currentVersionCode:"
-                            + currentVersionCode);
-                    if (serverVersionCode > currentVersionCode) {
-                        //有新版本了
-                        Log.d("Need to update");
-                        String url = list.get(0).apkUrl;
-                        String serverVersion = list.get(0).version;
-                        String changeLog = list.get(0).changeLog;
-                        //has new version
-                        Message message = new Message();
-                        Bundle bundle = new Bundle();
-                        bundle.putString("version", serverVersion);
-                        bundle.putInt("version_code", serverVersionCode);
-                        bundle.putString("url", url);
-                        bundle.putString("changeLog", changeLog);
-                        message.setData(bundle);
+            public void done(List<Version> list, BmobException e) {
+                if (e == null) {
+                    if (list != null && list.size() > 0) {
+                        int serverVersionCode = list.get(0).version_code;
+                        int currentVersionCode = BuildConfig.VERSION_CODE;
+                        Log.d("serverVersionCode:" + serverVersionCode + ",currentVersionCode:"
+                                + currentVersionCode);
+                        if (serverVersionCode > currentVersionCode) {
+                            //有新版本了
+                            Log.d("Need to update");
+                            String url = list.get(0).apkUrl;
+                            String serverVersion = list.get(0).version;
+                            String changeLog = list.get(0).changeLog;
+                            //has new version
+                            Message message = new Message();
+                            Bundle bundle = new Bundle();
+                            bundle.putString("version", serverVersion);
+                            bundle.putInt("version_code", serverVersionCode);
+                            bundle.putString("url", url);
+                            bundle.putString("changeLog", changeLog);
+                            message.setData(bundle);
 
-                        if (byUser) {
-                            listener.showUpdateNotification(serverVersion, url);
-                        } else {
-                            if (NetUtil.isWifiConnected(context)) {
-                                doDownloadNewVersionAPk(context, serverVersion, serverVersionCode,
-                                        changeLog,
-                                        url, listener);
+                            if (byUser) {
+                                listener.showUpdateNotification(serverVersion, url);
                             } else {
-                                //非wifi下，什么都不做了，浪费流量
+                                if (NetUtil.isWifiConnected(context)) {
+                                    doDownloadNewVersionAPk(context, serverVersion, serverVersionCode,
+                                            changeLog,
+                                            url, listener);
+                                } else {
+                                    //非wifi下，什么都不做了，浪费流量
 //                                listener.showUpdateNotification(serverVersion, url);
+                                }
+                            }
+                        } else {
+                            if (byUser) {
+                                listener.noVersionNeedToUpdate();
                             }
                         }
-                    } else {
-                        if (byUser) {
-                            listener.noVersionNeedToUpdate();
-                        }
                     }
+                } else {
+                    Log.e(e.getMessage());
                 }
-            }
-
-            @Override
-            public void onError(int i, String s) {
-                Log.e(s);
             }
         });
     }
 
     @Override
     public void uploadNewApk(final Context context, String filePath, final OnUploadListener listener) {
-        BmobProFile.getInstance(context).upload(filePath, new UploadListener() {
-            @Override
-            public void onSuccess(String s, String s1) {
-                Log.d("success.name:" + s + ",url:" + s1);
-                Version version =new Version();
-                version.version = BuildConfig.VERSION_NAME;
-                version.version_code = BuildConfig.VERSION_CODE;
-                version.apkUrl = s;
-                version.changeLog = context.getString(R.string.change_log);
-                version.update(context, "692ZQQQp", new UpdateListener() {
-                    @Override
-                    public void onSuccess() {
-                        Log.d();
-                        if (listener != null) {
-                            listener.onUploadSuccess();
-                        }
-                    }
 
-                    @Override
-                    public void onFailure(int i, String s) {
-                        Log.d();
-                        if (listener != null) {
-                            listener.onUploadFail(s);
-                        }
-                    }
-                });
-            }
-
+        final BmobFile bmobFile = new BmobFile(new File(filePath));
+        bmobFile.uploadblock(new UploadFileListener() {
             @Override
-            public void onProgress(int i) {
-                Log.d("progress：" + i);
-                if (listener != null) {
-                    listener.onUploadProgress(i);
+            public void done(BmobException e) {
+                if (e == null) {
+                    Log.d("success.name:" + bmobFile.getFilename() + ",url:" + bmobFile.getFileUrl());
+                    Version version =new Version();
+                    version.version = BuildConfig.VERSION_NAME;
+                    version.version_code = BuildConfig.VERSION_CODE;
+                    version.apkUrl = bmobFile.getFileUrl();
+                    version.changeLog = context.getString(R.string.change_log);
+                    version.update("692ZQQQp", new UpdateListener() {
+                        @Override
+                        public void done(BmobException e) {
+                            if (e == null) {
+                                Log.d();
+                                if (listener != null) {
+                                    listener.onUploadSuccess();
+                                }
+                            } else {
+                                Log.d(e.getMessage());
+                                if (listener != null) {
+                                    listener.onUploadFail(e.getMessage());
+                                }
+                            }
+                        }
+                    });
+                } else {
+                    Log.d("error:" + e.getMessage());
+                    if (listener != null) {
+                        listener.onUploadFail(e.getMessage());
+                    }
                 }
             }
 
             @Override
-            public void onError(int i, String s) {
-                Log.d("error:" + s);
+            public void onProgress(Integer value) {
+                super.onProgress(value);
+                Log.d("progress：" + value);
                 if (listener != null) {
-                    listener.onUploadFail(s);
+                    listener.onUploadProgress(value);
                 }
             }
         });
@@ -169,24 +169,24 @@ public class Main extends BaseMain implements IMain {
     private void doDownloadNewVersionAPk(final Context context, final String version, final int versionCode,
                                          final String chageLog,
                                          final String fileName, final OnCheckUpdateListener listener) {
-        BmobProFile.getInstance(context).download(fileName, new DownloadListener() {
+        BmobFile bmobFile = new BmobFile("test.apk", "", fileName);
+        bmobFile.download(new DownloadFileListener() {
             @Override
-            public void onSuccess(String s) {
-                Log.d("Download Apk Success.localPath:" + s);
-                SharedPreferencesUtil.put(context, "latest_version_code", versionCode);
-                SharedPreferencesUtil.put(context, "apkPath", s);
-                SharedPreferencesUtil.put(context, "changeLog", chageLog);
-                listener.onApkDownloaded(version, s, chageLog, true);
+            public void done(String s, BmobException e) {
+                if (e == null) {
+                    Log.d("Download Apk Success.localPath:" + s);
+                    SharedPreferencesUtil.put(context, "latest_version_code", versionCode);
+                    SharedPreferencesUtil.put(context, "apkPath", s);
+                    SharedPreferencesUtil.put(context, "changeLog", chageLog);
+                    listener.onApkDownloaded(version, s, chageLog, true);
+                } else {
+                    Log.e("Download apk faile:" + s);
+                }
             }
 
             @Override
-            public void onProgress(String s, int i) {
+            public void onProgress(Integer integer, long l) {
 
-            }
-
-            @Override
-            public void onError(int i, String s) {
-                Log.e("Download apk faile:" + s);
             }
         });
     }
@@ -196,17 +196,16 @@ public class Main extends BaseMain implements IMain {
         BmobCost bmobCost;
         for (final Cost cost : costList) {
             bmobCost = cost.getCostBean();
-            bmobCost.save(context, new SaveListener() {
+            bmobCost.save(new SaveListener<String>() {
                 @Override
-                public void onSuccess() {
-                    Log.d("upload success:" + cost.title);
-                    //上传成功后，删除本地数据
-                    cost.delete();
-                }
-
-                @Override
-                public void onFailure(int i, String s) {
-                    Log.d("upload failure:" + cost.title);
+                public void done(String objectId, BmobException e) {
+                    if (e == null) {
+                        Log.d("upload success:" + cost.title);
+                        //上传成功后，删除本地数据
+                        cost.delete();
+                    } else {
+                        Log.d("upload failure:" + cost.title);
+                    }
                 }
             });
         }
